Added deleting feed function (through long pressing the feed card) in FeedCard function. Refactored all functions from Cards.kt to MainActivity.kt and reorganized functions in logical order.
