+++ app/src/main/java/monster/minions/binocularss/activities/MainActivity.kt
@@ -1,26 +1,43 @@
 package monster.minions.binocularss.activities
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
+import android.graphics.Color
+import android.net.Uri
 import android.os.Bundle
 import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.animation.*
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.combinedClickable
+import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.graphics.ColorMatrix
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.core.content.ContextCompat
+import androidx.core.graphics.drawable.IconCompatParcelizer
 import androidx.lifecycle.*
 import androidx.navigation.NavController
 import androidx.navigation.NavHostController
@@ -30,12 +47,15 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import coil.annotation.ExperimentalCoilApi
+import coil.compose.rememberImagePainter
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.prof.rssparser.Parser
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.MutableStateFlow
+import monster.minions.binocularss.R
 import monster.minions.binocularss.activities.ui.theme.BinoculaRSSTheme
 import monster.minions.binocularss.dataclasses.Article
 import monster.minions.binocularss.dataclasses.Feed
@@ -133,52 +153,6 @@ class MainActivity : ComponentActivity() {
     }
 
     /**
-     * Replace the unmodified article with a modified article.
-     * This is to be used when updating article.bookmarked and article.read
-     *
-     * @param modifiedArticle Article with a modified property.
-     */
-    private fun setArticle(modifiedArticle: Article, refreshBookmark: Boolean = true) {
-        for (feed in feedGroup.feeds) {
-            val articles = feed.articles.toMutableList()
-            for (unmodifiedArticle in articles) {
-                if (modifiedArticle == unmodifiedArticle) {
-                    feed.articles.remove(unmodifiedArticle)
-                    feed.articles.add(modifiedArticle)
-                    break
-                }
-            }
-        }
-
-        articleList.value = sortArticlesByDate(getAllArticles(feedGroup))
-
-        if (refreshBookmark) {
-            bookmarkedArticleList.value = sortArticlesByDate(getBookmarkedArticles(feedGroup))
-        }
-
-        feedList.value = sortFeedsByTitle(feedGroup.feeds)
-    }
-
-    /**
-     * Replace the unmodified article with a modified article.
-     * This is to be used when updating feed.tags
-     *
-     * @param modifiedFeed Feed with a modified property.
-     */
-    private fun setFeeds(modifiedFeed: Feed) {
-        val feeds = feedGroup.feeds.toMutableList()
-        for (unmodifiedFeed in feeds) {
-            if (modifiedFeed == unmodifiedFeed) {
-                feedGroup.feeds.remove(unmodifiedFeed)
-                feedGroup.feeds.add(modifiedFeed)
-                break
-            }
-        }
-
-        feedList.value = sortFeedsByTitle(feedGroup.feeds)
-    }
-
-    /**
      * Save the list of user feeds to the Room database (feed-db) for data persistence.
      *
      * The database files can be found in `Android/data/data/monster.minions.binocularss.databases`.
@@ -219,6 +193,259 @@ class MainActivity : ComponentActivity() {
         articleList.value = sortArticlesByDate(getAllArticles(feedGroup))
         bookmarkedArticleList.value = sortArticlesByDate(getBookmarkedArticles(feedGroup))
         feedList.value = sortFeedsByTitle(feedGroup.feeds)
+    }
+
+    /**
+     * Replace the unmodified article with a modified article.
+     * This is to be used when updating article.bookmarked and article.read
+     *
+     * @param modifiedArticle Article with a modified property.
+     */
+    private fun setArticle(modifiedArticle: Article, refreshBookmark: Boolean = true) {
+        for (feed in feedGroup.feeds) {
+            val articles = feed.articles.toMutableList()
+            for (unmodifiedArticle in articles) {
+                if (modifiedArticle == unmodifiedArticle) {
+                    feed.articles.remove(unmodifiedArticle)
+                    feed.articles.add(modifiedArticle)
+                    break
+                }
+            }
+        }
+
+        articleList.value = sortArticlesByDate(getAllArticles(feedGroup))
+
+        if (refreshBookmark) {
+            bookmarkedArticleList.value = sortArticlesByDate(getBookmarkedArticles(feedGroup))
+        }
+
+        feedList.value = sortFeedsByTitle(feedGroup.feeds)
+    }
+
+    /**
+     * Replace the unmodified feeds with a modified feed.
+     * This is to be used when updating feed.tags
+     *
+     * @param modifiedFeed Feed with a modified property.
+     */
+    private fun setFeeds(modifiedFeed: Feed) {
+        val feeds = feedGroup.feeds.toMutableList()
+        for (unmodifiedFeed in feeds) {
+            if (modifiedFeed == unmodifiedFeed) {
+                feedGroup.feeds.remove(unmodifiedFeed)
+                feedGroup.feeds.add(modifiedFeed)
+                break
+            }
+        }
+
+        feedList.value = sortFeedsByTitle(feedGroup.feeds)
+    }
+
+    /**
+     * Delete feeds through deleting the source from feedDao
+     *
+     * @param feed A feed that the user wants to delete
+     */
+    fun deleteFeed(feed: Feed) {
+
+        feedDao.deleteBySource(feed.source)
+        feedGroup.feeds.remove(feed)
+
+        articleList.value = mutableListOf()
+        feedList.value = mutableListOf()
+        bookmarkedArticleList.value = mutableListOf()
+    }
+
+    /**
+     * Formats the image and article/feed description if available.
+     * If not, put in a placeholder image (tminions logo) or empty string for description
+     *
+     * @param image string that represents the URL of the image
+     * @param description Article/Feed descriptions
+     */
+    @Composable
+    fun CardImage(image: String, description: String = "") {
+
+        // Color matrix to turn image grayscale
+        val grayScaleMatrix = ColorMatrix(
+            floatArrayOf(
+                0.33f, 0.33f, 0.33f, 0f, 0f,
+                0.33f, 0.33f, 0.33f, 0f, 0f,
+                0.33f, 0.33f, 0.33f, 0f, 0f,
+                0f, 0f, 0f, 1f, 0f
+            )
+        )
+
+        val imageExists = image != "null" && image != null
+
+        // Box for image on the right.
+        Card(
+            modifier = Modifier.fillMaxWidth(),
+            shape = RoundedCornerShape(10.dp),
+            elevation = 4.dp
+        ) {
+            Box(
+                modifier = Modifier
+                    .size(150.dp, 150.dp)
+                    .background(MaterialTheme.colors.background, RoundedCornerShape(4.dp))
+            ) {
+                Image(
+                    painter = rememberImagePainter(
+                        data =
+                        if (imageExists) image
+                        else "https://avatars.githubusercontent.com/u/91392435?s=200&v=4",
+                        builder = {
+                            // Placeholder when the image hasn't loaded yet.
+                            placeholder(R.drawable.ic_launcher_foreground)
+                        }
+                    ),
+                    contentScale = ContentScale.Crop,
+                    contentDescription = description,
+                    colorFilter = if (imageExists) null else ColorFilter.colorMatrix(
+                        grayScaleMatrix
+                    ),
+                    modifier = Modifier
+                        .fillMaxSize()
+                )
+            }
+        }
+    }
+
+    /**
+     * Displays a single feed in a card view format
+     * Includes a long hold function to delete a feed
+     *
+     * @param context \\TODO fill this in
+     * @param feed The feed to be displayed
+     */
+    @ExperimentalCoilApi
+    @Composable
+    fun FeedCard(context: Context, feed: Feed) {
+        var showDropdown by remember { mutableStateOf(false) }
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp)
+                .pointerInput(Unit) {
+                    detectTapGestures(
+                        onLongPress = { showDropdown = true },
+                        onTap = {
+                            // TODO temporary until articleFromFeed
+                            val intent = Intent(Intent.ACTION_VIEW)
+                            intent.data = Uri.parse(feed.link)
+                            ContextCompat.startActivity(context, intent, null)
+                        }
+                    )
+                }, elevation = 4.dp
+        ) {
+            Column {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ) {
+                    // Column for feed title.
+                    Column(
+                        modifier = Modifier.width(200.dp)
+                    ) {
+                        feed.title?.let { title ->
+                            Text(text = title, fontWeight = FontWeight.SemiBold)
+                        }
+                        val items = listOf("Delete")
+                        // TODO show up wherever we longpress
+                        DropdownMenu(
+                            expanded = showDropdown,
+                            onDismissRequest = { showDropdown = false },
+                            modifier = Modifier
+                                .background(MaterialTheme.colors.background)
+                        ) {
+                            items.forEach { item ->
+                                DropdownMenuItem(onClick = {
+                                    when (item) {
+                                        "Delete" -> {
+                                            deleteFeed(feed)
+                                        }
+                                    }
+                                }) {
+                                    Text(text = item)
+                                }
+                            }
+                        }
+                    }
+                    val imageExists = feed.image != "null"
+
+                    CardImage(image = feed.image, description = feed.description!!)
+                }
+            }
+
+            // Row for buttons in the future that is currently not used
+            // Row(
+            //     modifier = Modifier
+            //         .fillMaxWidth()
+            // ) {
+            // }
+        }
+    }
+
+    /**
+     * Displays a single article in a card view format
+     *
+     * @param article The article to be displayed
+     */
+    @SuppressLint("SimpleDateFormat")
+    @ExperimentalCoilApi
+    @Composable
+    fun ArticleCard(context: Context, article: Article, updateValues: (article: Article) -> Unit) {
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp)
+                .clickable {
+                    val intent = Intent(context, ArticleActivity::class.java)
+                    intent.putExtra("article", article)
+                    article.read = true
+                    updateValues(article)
+                    ContextCompat.startActivity(context, intent, null)
+                },
+            elevation = 4.dp
+        ) {
+            Column {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ) {
+                    // Column for title, feed, and time.
+                    Column(
+                        modifier = Modifier
+                            .width(200.dp)
+                            .padding(end = 12.dp)
+                    ) {
+                        article.title?.let { title ->
+                            Text(text = title, fontWeight = FontWeight.SemiBold)
+                        }
+                        Text(text = article.sourceTitle)
+                        Text(text = getTime(article.pubDate!!))
+                    }
+
+                    CardImage(image = article.image!!, description = article.description!!)
+                }
+
+                // Row for buttons on the bottom.
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceEvenly
+                ) {
+                    BookmarkFlag(article = article) { updateValues(article) }
+                    ShareFlag(context = context, article = article)
+                    ReadFlag(article = article) { updateValues(article) }
+                    BrowserFlag(context = context, article = article)
+                }
+            }
+        }
     }
 
     /**
@@ -264,23 +491,6 @@ class MainActivity : ComponentActivity() {
     }
 
     /**
-     * Displays the list of articles associated with a given feed
-     *
-     * TODO to be implemented
-     */
-    @Composable
-    fun ArticlesFromFeed() {
-        LazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            contentPadding = PaddingValues(bottom = 80.dp)
-        ) {
-            items(items = currentFeed.articles) { article ->
-                ArticleCard(context = this@MainActivity, article = article) { setArticle(it) }
-            }
-        }
-    }
-
-    /**
      * Displays a list of articles in the order given by the currently selected sorting method
      */
     @Composable
@@ -320,6 +530,23 @@ class MainActivity : ComponentActivity() {
                 items(items = articles) { article ->
                     ArticleCard(context = this@MainActivity, article = article) { setArticle(it) }
                 }
+            }
+        }
+    }
+
+    /**
+     * Displays the list of articles associated with a given feed
+     *
+     * TODO to be implemented
+     */
+    @Composable
+    fun ArticlesFromFeed() {
+        LazyColumn(
+            modifier = Modifier.fillMaxSize(),
+            contentPadding = PaddingValues(bottom = 80.dp)
+        ) {
+            items(items = currentFeed.articles) { article ->
+                ArticleCard(context = this@MainActivity, article = article) { setArticle(it) }
             }
         }
     }
@@ -465,7 +692,7 @@ class MainActivity : ComponentActivity() {
     }
 
     /**
-     * Bottom app bar with buttons for article and feed views.
+     * Bottom app bar with buttons for article, feed, and bookmark views.
      */
     @Composable
     fun BottomNavigationBar(navController: NavController) {
@@ -512,22 +739,6 @@ class MainActivity : ComponentActivity() {
         }
     }
 
-    /**
-     * Animate the entrance animation for navigation items.
-     */
-    @ExperimentalAnimationApi
-    @Composable
-    fun EnterAnimation(content: @Composable () -> Unit) {
-        // TODO figure out non-deprecated library
-        AnimatedVisibility(
-            visible = true,
-            enter = slideInVertically(initialOffsetY = { 10000 }),
-            exit = fadeOut(),
-            initiallyVisible = false
-        ) {
-            content()
-        }
-    }
 
     /**
      * Composable that loads in and out views based on the current navigation item selected.
@@ -551,6 +762,23 @@ class MainActivity : ComponentActivity() {
                     BookmarksView(navController)
                 }
             }
+        }
+    }
+
+    /**
+     * Animate the entrance animation for navigation items.
+     */
+    @ExperimentalAnimationApi
+    @Composable
+    fun EnterAnimation(content: @Composable () -> Unit) {
+        // TODO figure out non-deprecated library
+        AnimatedVisibility(
+            visible = true,
+            enter = slideInVertically(initialOffsetY = { 10000 }),
+            exit = fadeOut(),
+            initiallyVisible = false
+        ) {
+            content()
         }
     }
 
@@ -609,3 +837,30 @@ class MainActivity : ComponentActivity() {
         UI()
     }
 }
+
+
+// Incorporated into FeedCard.
+//    @Composable
+//    fun FeedsDropDownMenu(
+//        feed: Feed,
+//        expanded: Boolean,
+//        onSelect: (selectedItem: String, feed: Feed) -> Unit
+//    ) {
+//        var showDropdown = expanded
+//        val items = listOf("Delete")
+//        DropdownMenu(
+//            expanded = showDropdown,
+//            onDismissRequest = { showDropdown = false },
+//            modifier = Modifier
+//                .fillMaxWidth()
+//                .background(MaterialTheme.colors.background)
+//        ) {
+//            items.forEach { item ->
+//                DropdownMenuItem(onClick = {
+//                    onSelect(item, feed)
+//                }) {
+//                    Text(text = item)
+//                }
+//            }
+//        }
+//    }
